+++ docs/whats_new.rst
@@ -42,6 +42,9 @@ Enhancements
 
 New regions
 ~~~~~~~~~~~
+- Added :py:attr:`natural_earth.land_10` and :py:attr:`natural_earth.land_50` regions from
+  natural earth (:pull:`195`) by `Martin van Driel <https://github.com/martinvandriel>`_.
+
 
 Bug Fixes
 ~~~~~~~~~